+++ AtdIT/src/main/java/persistence/databaseManager/DatabaseManager.java
@@ -1,4 +1,4 @@
-package main.java.persistence;
+package main.java.persistence.databaseManager;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
