+++ AtdIT/src/main/java/persistence/databaseManager/DatabaseManagerInterface.java
@@ -1,4 +1,4 @@
-package main.java.persistence;
+package main.java.persistence.databaseManager;
 
 import main.java.persistence.databaseTables.AbstractEntry;
